e2e: swap WebDriver to Firefox/Gecko instead
As only Firefox is available on the Travis CI testing
environment, the e2e test should not default to Chrome.

In order to support Gecko (especially later versions),
it is beneficial to use Selenium 3, so that the older
FirefoxDriver does not take precedence over GeckoDriver.

Thus, this patch includes the following changes:
	- Default Selenium env set to 'firefox'
	- Required Selenium version is now 3.0.1
	- Addition of 'geckodriver' dependency

Signed-off-by: Syed Abdullah <contact@syedabdullah.net>